requirements.txt
# Core application
fastapi==0.115.12
uvicorn==0.34.3
uvloop==0.21.0
httpx==0.27.0
loguru==0.7.2

# Settings and .env loading
pydantic==2.8.2
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Database and Data
SQLAlchemy==2.0.41
psycopg==3.1.18
pandas==2.3.0

# Docker interaction
docker==7.1.0
psycopg2-binary